generate-password/requirements.txt
psycopg2-binary
pyotp
qrcode
pillow

